Did not know page.tsx cannot be client component
